Add `perfect_power()` to the public API
