+++ galois/modular.py
@@ -6,11 +6,10 @@ import numpy as np
 from .factor import factors
 from .math_ import lcm, prod
 from .overrides import set_module
-from .prime import is_prime
 
 __all__ = [
     "gcd", "egcd", "are_coprime", "crt",
-    "totatives", "euler_phi", "carmichael_lambda", "legendre_symbol", "jacobi_symbol", "kronecker_symbol", "is_cyclic",
+    "totatives", "euler_phi", "carmichael_lambda", "is_cyclic",
     "is_primitive_root", "primitive_root", "primitive_roots",
 ]
 
@@ -364,192 +363,6 @@ def carmichael_lambda(n):
         lambdas.append(l)
 
     return lcm(*lambdas)
-
-
-@set_module("galois")
-def legendre_symbol(a, p):
-    r"""
-    Computes the Legendre symbol :math:`(\frac{a}{p})`.
-
-    The Legendre symbol is useful for determining if :math:`a` is a quadratic residue modulo :math:`p`, namely
-    :math:`a \in Q_p`. A quadratic residue :math:`a` modulo :math:`p` satisfies :math:`x^2 \equiv a\ (\textrm{mod}\ p)`
-    for some :math:`x`.
-
-    .. math::
-
-        \bigg(\frac{a}{p}\bigg) =
-            \begin{cases}
-                0, & p\ |\ a
-
-                1, & a \in Q_p
-
-                -1, & a \in \overline{Q}_p
-            \end{cases}
-
-    Parameters
-    ----------
-    a : int
-        An integer.
-    p : int
-        An odd prime :math:`p \ge 3`.
-
-    Returns
-    -------
-    int
-        The Legendre symbol :math:`(\frac{a}{p})` with value in :math:`\{0, 1, -1\}`.
-
-    References
-    ----------
-    * Algorithm 2.149 from https://cacr.uwaterloo.ca/hac/about/chap2.pdf
-
-    Examples
-    --------
-    The quadratic residues modulo :math:`7` are :math:`Q_7 = \{1, 2, 4\}`. The quadratic non-residues
-    modulo :math:`7` are :math:`\overline{Q}_7 = \{3, 5, 6\}`.
-
-    .. ipython:: python
-
-        [pow(x, 2, 7) for x in range(7)]
-        for a in range(7):
-            print(f"({a} / 7) = {galois.legendre_symbol(a, 7)}")
-    """
-    if not is_prime(p):
-        raise ValueError(f"Argument `p` must be an odd prime greater than 2, not {p}.")
-
-    return jacobi_symbol(a, p)
-
-
-@set_module("galois")
-def jacobi_symbol(a, n):
-    r"""
-    Computes the Jacobi symbol :math:`(\frac{a}{n})`.
-
-    The Jacobi symbol extends the Legendre symbol for odd :math:`n \ge 3`. Unlike the Legendre symbol, :math:`(\frac{a}{n}) = 1`
-    does not imply :math:`a` is a quadratic residue modulo :math:`n`. However, all :math:`a \in Q_n` have :math:`(\frac{a}{n}) = 1`.
-
-    Parameters
-    ----------
-    a : int
-        An integer.
-    n : int
-        An odd integer :math:`n \ge 3`.
-
-    Returns
-    -------
-    int
-        The Jacobi symbol :math:`(\frac{a}{n})` with value in :math:`\{0, 1, -1\}`.
-
-    References
-    ----------
-    * Algorithm 2.149 from https://cacr.uwaterloo.ca/hac/about/chap2.pdf
-
-    Examples
-    --------
-    The quadratic residues modulo :math:`9` are :math:`Q_9 = \{1, 4, 7\}` and these all satisfy :math:`(\frac{a}{9}) = 1`.
-    The quadratic non-residues modulo :math:`9` are :math:`\overline{Q}_9 = \{2, 3, 5, 6, 8\}`, but notice :math:`\{2, 5, 8\}`
-    also satisfy :math:`(\frac{a}{9}) = 1`. The set of integers :math:`\{3, 6\}` not coprime to :math:`n` satisfies
-    :math:`(\frac{a}{9}) = 0`.
-
-    .. ipython:: python
-
-        [pow(x, 2, 9) for x in range(9)]
-        for a in range(9):
-            print(f"({a} / 9) = {galois.jacobi_symbol(a, 9)}")
-    """
-    if not isinstance(a, (int, np.integer)):
-        raise TypeError(f"Argument `a` must be an integer, not {type(a)}.")
-    if not isinstance(n, (int, np.integer)):
-        raise TypeError(f"Argument `n` must be an integer, not {type(n)}.")
-    if not (n > 2 and n % 2 == 1):
-        raise ValueError(f"Argument `n` must be an odd integer greater than 2, not {n}.")
-
-    a = a % n
-    if a == 0:
-        return 0
-    if a == 1:
-        return 1
-
-    # Write a = 2^e * a1
-    a1, e = a, 0
-    while a1 % 2 == 0:
-        a1, e = a1 // 2, e + 1
-    assert 2**e * a1 == a
-
-    if e % 2 == 0:
-        s = 1
-    else:
-        if n % 8 in [1, 7]:
-            s = 1
-        else:
-            s = -1
-
-    if n % 4 == 3 and a1 % 4 == 3:
-        s = -s
-
-    n1 = n % a1
-    if a1 == 1:
-        return s
-    else:
-        return s * jacobi_symbol(n1, a1)
-
-
-@set_module("galois")
-def kronecker_symbol(a, n):
-    r"""
-    Computes the Kronecker symbol :math:`(\frac{a}{n})`.
-
-    The Kronecker symbol extends the Jacobi symbol for all :math:`n`.
-
-    Parameters
-    ----------
-    a : int
-        An integer.
-    n : int
-        An integer.
-
-    Returns
-    -------
-    int
-        The Kronecker symbol :math:`(\frac{a}{n})` with value in :math:`\{0, -1, 1\}`.
-
-    References
-    ----------
-    * Algorithm 2.149 from https://cacr.uwaterloo.ca/hac/about/chap2.pdf
-    """
-    # pylint: disable=too-many-return-statements
-    if not isinstance(a, (int, np.integer)):
-        raise TypeError(f"Argument `a` must be an integer, not {type(a)}.")
-    if not isinstance(n, (int, np.integer)):
-        raise TypeError(f"Argument `n` must be an integer, not {type(n)}.")
-
-    if n == 0:
-        return 1 if a in [1, -1] else 0
-    if n == 1:
-        return 1
-    if n == -1:
-        return -1 if a < 0 else 1
-    if n == 2:
-        if a % 2 == 0:
-            return 0
-        elif a % 8 in [1, 7]:
-            return 1
-        else:
-            return -1
-
-    # Factor out the unit +/- 1
-    u = -1 if n < 0 else 1
-    n //= u
-
-    # Factor out the powers of 2 so the resulting n is odd
-    e = 0
-    while n % 2 == 0:
-        n, e = n // 2, e + 1
-
-    if n >=3 :
-        # Handle the remaining odd n using the Jacobi symbol
-        return kronecker_symbol(a, u) * kronecker_symbol(a, 2)**e * jacobi_symbol(a, n)
-    else:
-        return kronecker_symbol(a, u) * kronecker_symbol(a, 2)**e
 
 
 @set_module("galois")
